Fix lane switching bugs
Due to a number of factors agents got incorrect information about
whether they could switch lanes, causing them to never do so.
These should all be fixed, but because of the finicky nature of the this
whole subsystem it could be nice to add some tests (time permitting of
course).